Split code into 2 files #10
